requirements.txt
pillow==9.5.0
joblib==1.3.1
pyyaml==6.0.1
opencv-python==4.8.0.74
scikit-learn==1.3.0
matplotlib==3.7.1 
tqdm==4.65.0
jsonargparse==4.22.1
numpy==1.24.2
torchvision==0.15.1
wandb